backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pandas
numpy
networkx
scipy
python-dotenv